requirements.txt
uvicorn
pytest-asyncio
Unidecode==1.3.8
pydantic-settings==2.2.0
fastapi==0.109.2
dependency-injector
python-multipart
alembic==1.13.1
asyncpg==0.29.0
httpx==0.28.1
SQLAlchemy==2.0.27
pyjwt==2.8.0